# Add abstraction layer to support multiple algorithms