--- a/include/engine/algorithm.hpp
+++ b/include/engine/algorithm.hpp
@@ -18,6 +18,10 @@ struct CH final
 struct CoreCH final
 {
 };
+
+template<typename AlgorithmT> const char* name();
+template<> inline const char* name<CH>() { return "CH"; }
+template<> inline const char* name<CoreCH>() { return "CoreCH"; }
 }
 
 namespace algorithm_trais