--- a/include/engine/algorithm.hpp
+++ b/include/engine/algorithm.hpp
@@ -19,9 +19,9 @@ struct CoreCH final
 {
 };
 
-template<typename AlgorithmT> const char* name();
-template<> inline const char* name<CH>() { return "CH"; }
-template<> inline const char* name<CoreCH>() { return "CoreCH"; }
+template <typename AlgorithmT> const char *name();
+template <> inline const char *name<CH>() { return "CH"; }
+template <> inline const char *name<CoreCH>() { return "CoreCH"; }
 }
 
 namespace algorithm_trais