--- a/include/engine/data_watchdog.hpp
+++ b/include/engine/data_watchdog.hpp
@@ -24,9 +24,10 @@ namespace engine
 // This class monitors the shared memory region that contains the pointers to
 // the data and layout regions that should be used. This region is updated
 // once a new dataset arrives.
-class DataWatchdog
+template <typename AlgorithmT> class DataWatchdog final
 {
     using mutex_type = typename storage::SharedMonitor<storage::SharedDataTimestamp>::mutex_type;
+    using FacadeT = datafacade::ContiguousInternalMemoryDataFacade<AlgorithmT>;
 
   public:
     DataWatchdog() : active(true), timestamp(0)
@@ -35,7 +36,7 @@ class DataWatchdog
         {
             boost::interprocess::scoped_lock<mutex_type> current_region_lock(barrier.get_mutex());
 
-            facade = std::make_shared<datafacade::ContiguousInternalMemoryDataFacade>(
+            facade = std::make_shared<const FacadeT>(
                 std::make_unique<datafacade::SharedMemoryAllocator>(barrier.data().region));
             timestamp = barrier.data().timestamp;
         }
@@ -43,14 +44,14 @@ class DataWatchdog
         watcher = std::thread(&DataWatchdog::Run, this);
     }
 
-    ~DataWatchdog()
+    virtual ~DataWatchdog()
     {
         active = false;
         barrier.notify_all();
         watcher.join();
     }
 
-    auto GetDataFacade() const { return facade; }
+    std::shared_ptr<const FacadeT> Get() const { return facade; }
 
   private:
     void Run()
@@ -67,7 +68,7 @@ class DataWatchdog
             if (timestamp != barrier.data().timestamp)
             {
                 auto region = barrier.data().region;
-                facade = std::make_shared<datafacade::ContiguousInternalMemoryDataFacade>(
+                facade = std::make_shared<const FacadeT>(
                     std::make_unique<datafacade::SharedMemoryAllocator>(region));
                 timestamp = barrier.data().timestamp;
                 util::Log() << "updated facade to region " << region << " with timestamp "
@@ -82,7 +83,7 @@ class DataWatchdog
     std::thread watcher;
     bool active;
     unsigned timestamp;
-    std::shared_ptr<datafacade::ContiguousInternalMemoryDataFacade> facade;
+    std::shared_ptr<const FacadeT> facade;
 };
 }
 }