… include/engine/edge_unpacker.hpp
… include/engine/plugins/match.hpp
… include/engine/plugins/nearest.hpp
… include/engine/plugins/plugin_base.hpp
… include/engine/plugins/table.hpp
… include/engine/plugins/tile.hpp
… include/engine/plugins/trip.hpp
… include/engine/plugins/viaroute.hpp
… include/engine/routing_algorithms/alternative_path.hpp
… include/engine/routing_algorithms/direct_shortest_path.hpp
… include/engine/routing_algorithms/many_to_many.hpp
… include/engine/routing_algorithms/map_matching.hpp
… include/engine/routing_algorithms/routing_base.hpp
… include/engine/routing_algorithms/shortest_path.hpp
… include/engine/routing_algorithms/tile_turns.hpp
… include/engine/trip/trip_brute_force.hpp
… include/engine/trip/trip_farthest_insertion.hpp
… include/osrm/osrm.hpp
… include/osrm/osrm_fwd.hpp
… include/storage/shared_datatype.hpp
… include/util/binary_heap.hpp
… src/engine/engine.cpp
… src/engine/hint.cpp
… src/engine/plugins/match.cpp
… src/engine/plugins/nearest.cpp
… src/engine/plugins/table.cpp
… src/engine/plugins/tile.cpp
… src/engine/plugins/trip.cpp
… src/engine/plugins/viaroute.cpp
… src/engine/routing_algorithms/alternative_path.cpp