# Issue

This implements #3208 by using a template tag to automatically generate a class hierarchy for an algorithm and only use actual virtual function calls in a minimum of places.

This design is the result of the following prototype https://gist.github.com/TheMarex/8ba7594cab6082f5c39f08e79b0152db that compares a manually written OOM scheme against template based code.

## Compile time

    ccache -C
    cmake ..
    time make osrm # cold build

`master`

   make osrm  104.80s user 7.58s system 99% cpu 1:52.99 total

`refactor/algorithm-aware`

   make osrm  121.77s user 8.10s system 99% cpu 2:10.49 total

## Binary size

    ls -l libosrm.a

`master`

    6052202

`refactor/algorithm-aware`

    6590768

## Performance

Measuring the 10000 random requests on California the result as basically the same:

`master`

median: 0.462845 p95: 11.4442 p05: 0.322627
median: 0.459129 p95: 11.2884 p05: 0.320714
median: 0.498818 p95: 11.7293 p05: 0.339625

`refactor/algorithm-aware`

median: 0.462679 p95: 11.3626 p05: 0.319035
median: 0.462412 p95: 11.2776 p05: 0.321605
median: 0.456968 p95: 11.3829 p05: 0.32047

## Tasklist
 - [x] Finish up #3528 
 - [x] micro-benchmarks for osrm-routed
 - [x] review
 - [x] adjust for comments

## Requirements / Relations

Depends on #3528 to remove the query locks for `SharedMemoryDataFacade`.