@TheMarex 2017-02-24
@oxidase I would prefer two separate PRs. This PR is not supposed to change/add any functionality to make it easier to find regressions.