@TheMarex 2017-02-24
I think I found a solution to abstract away the algorithms for the plugins so we don't need them in the header file. Will try this first before we start the review process.